refactor(auth): restructure authentication logic and types for improved maintainability
