+++ MODULE_ANALYSIS.md
@@ -9,7 +9,7 @@ The current architecture has a decent separation of concerns, but there are area
 ### Detected Issues
 
 1.  **Improper Separation of Concerns**:
-    *   **`AuthContext` Overload**: The `AuthContext` (`src/contexts/auth-context.tsx`) handles too many responsibilities. It manages authentication state, API calls for login and password reset, token management, and routing. This entangles UI logic (routing), business logic (authentication), and network logic (API calls).
+    *   **`AuthContext` Overload**: The `AuthContext` (`src/contexts/auth-context.tsx`) handled too many responsibilities. It managed authentication state, API calls for login and password reset, token management, and routing. This has been refactored by moving `AuthContext` to `src/providers/auth-provider.tsx` and extracting API calls and types to `src/features/auth/api` and `src/features/auth/types` respectively.
     *   **Component-Level Fetching**: Components like `LoginPage` and `ForgotPasswordPage` contain logic for handling user input and form submission, but the actual API calls are abstracted away in `AuthContext`. This is a good first step, but the context itself is doing too much.
     *   **`page.tsx` as a Controller**: The main `page.tsx` acts as a controller, deciding which module to render based on the application state. It also contains hardcoded data for the dashboard. This mixes presentation logic with data provisioning.
 
@@ -59,19 +59,14 @@ To improve scalability and maintainability, I propose a feature-based folder str
 
 Based on the suggested folder restructure, here are some specific code refactoring suggestions:
 
-1.  **Refactor `AuthContext`**:
-    *   **Move API calls**: Move the `loginApi`, `resetPasswordApi`, and `validateTokenApi` calls from `AuthContext` to `src/features/auth/api/index.ts`.
-    *   **Create `useAuth` hook**: Create a custom hook `src/features/auth/hooks/useAuth.ts` that encapsulates the logic for interacting with the `AuthContext` and making API calls. The components will use this hook instead of directly interacting with the context.
-    *   **Simplify `AuthProvider`**: The `AuthProvider` in `src/providers/auth-provider.tsx` should only be responsible for managing the authentication state (user, token, isAuthenticated). The business logic should be in the `useAuth` hook.
-
-2.  **Decouple `page.tsx`**:
+1.  **Decouple `page.tsx`**:
     *   **Use a routing solution**: Instead of a giant `switch` statement, use a more dynamic routing solution. The Next.js app router already handles this. The main `page.tsx` should be a simple layout, and the feature modules should be rendered as separate pages.
     *   **Fetch data in components**: The dashboard data should be fetched within the `DashboardModule` component, not in the main `page.tsx`.
 
-3.  **Organize API Calls**:
+2.  **Organize API Calls**:
     *   Create separate API files for each feature under `src/features/<feature>/api`. For example, `src/features/inventory/api/index.ts` would contain all API calls related to inventory.
 
-4.  **Co-locate Types**:
+3.  **Co-locate Types**:
     *   Move the type definitions from `src/lib/types` to the feature folders they belong to. For example, `User` and `Auth` related types should be in `src/features/auth/types/index.ts`.
 
 By implementing these changes, the WMS-FE application will have a more robust and scalable architecture, making it easier to maintain and extend in the future.
