+++ src/contexts/auth-context.test.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { render, act, waitFor } from "@testing-library/react";
 import { AuthProvider, AuthContextType } from "./auth-context";
-import { User } from "@/lib/types";
 
 // Mocks
 const mockPush = jest.fn();
@@ -49,6 +48,7 @@ const localStorageMock = (() => {
 Object.defineProperty(window, "localStorage", { value: localStorageMock });
 
 import { useAuth } from "@/features/auth/hooks/useAuth";
+import { User } from "@/features/auth/types";
 
 function TestComponent() {
   const ctx = useAuth();
