+++ src/lib/types/index.ts
@@ -15,7 +15,7 @@ export * from './dock.types';
 export * from './order.types';
 export * from './inventory.types';
 export * from './customer.types';
-export * from './user.types';
+
 export * from './filters.types';
 export * from './report.types';
 export * from './return.types';
